+++ Assets/Plugins/CircleScroller/Samples~/CircleScrollerExample/Scripts/ExampleScene.cs
@@ -11,7 +11,7 @@ public class ExampleScene : MonoBehaviour
     {
         for (int k = 0; k < this.circleScroller.Length; k++)
         {
-            List<CircleButtonExampleData> circleButtonDatas = new List<CircleButtonExampleData>();
+            this.circleScroller[k].Clear();
 
             for (int i = 0; i < 8; i++)
             {
@@ -23,10 +23,10 @@ public class ExampleScene : MonoBehaviour
                     Debug.Log(circleButtonData.text);
                 };
 
-                circleButtonDatas.Add(circleButtonData);
+                this.circleScroller[k].Add(circleButtonData);
             }
 
-            this.circleScroller[k].Init(circleButtonDatas);
+            this.circleScroller[k].Refresh();
             this.circleScroller[k].SnapToCenter(1);
         }
     }