+++ Assets/Plugins/CircleScroller/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## CHANGELOG
 
+### v1.1.0
+- Optimized code in CircleScroller, it's now more clearer for user to use with Add/Remove/Clear/Refresh method.
+- Example provided in project is also modified.
+
 ### v1.0.0
 - A plugin that allows you create gui button with circle layout.